updates based on review
