+++ integration/ssm-dynamic/package.json
@@ -4,11 +4,11 @@
         "@types/node": "^10.0.0"
     },
     "dependencies": {
-        "@pulumi/aws": "^6.0.0",
+        "@pulumi/aws": "6.61.0",
         "@pulumi/aws-native": "^1.11.0",
-        "@pulumi/cdk": "^0.5.0",
+        "@pulumi/cdk": "^1.0.0",
         "@pulumi/pulumi": "^3.0.0",
-        "aws-cdk-lib": "2.149.0",
+        "aws-cdk-lib": "2.156.0",
         "constructs": "10.3.0",
         "esbuild": "^0.24.0"
     }
